resolved pr comments and merged some of the render files together
