+++ pages/upload.tsx
@@ -16,10 +16,8 @@ import SchemaSelector from '@/src/Form/SchemaSelector'
 import SubmissionError from '@/src/Form/SubmissionError'
 import Form from '@/src/Form/Form'
 import ModelExportAndSubmission from '@/src/Form/ModelExportAndSubmission'
-import RenderFileTab, { FileTabComplete } from '@/src/Form/RenderFileTab'
-import RenderBasicFileTab from '@/src/Form/RenderBasicFileTab'
-import RenderBuildOptionsTab from '@/src/Form/RenderBuildOptionsTab'
-import RenderBasicBuildOptionsTab from '@/src/Form/RenderBasicBuildOptionsTab'
+import { RenderFileTab, RenderBasicFileTab, FileTabComplete } from '@/src/Form/RenderFileTab'
+import { RenderBuildOptionsTab, RenderBasicBuildOptionsTab } from '@/src/Form/RenderBuildOptionsTab'
 import { useGetCurrentUser } from '@/data/user'
 import { MinimalErrorWrapper } from '@/src/errors/ErrorWrapper'
 import LoadingBar from '@/src/common/LoadingBar'
