+++ types/interfaces.ts
@@ -1,8 +1,6 @@
 import { Date, Types } from 'mongoose'
 import Logger from 'bunyan'
 import { UserDoc } from '../server/models/User'
-import { XYPosition } from 'react-flow-renderer'
-import { boolean } from 'yargs'
 
 export type { VersionDoc as Version } from '../server/models/Version'
 export type { DeploymentDoc as Deployment } from '../server/models/Deployment'
